read(...,subscribe=) >0 interpreted as float timeout, <0 blocking
git-svn-id: http://sbfsvn01.cl.gemini.edu/software/yorick-svipc/trunk@24247 ee6ba543-07f8-4773-b02c-01d50e9d00ba

+++ common/svipc.h
@@ -50,7 +50,7 @@ long svipc_ftok(char *path, int proj);
 int  svipc_shm_info(long key, long details);
 int  svipc_shm_init(long key, long numslots);
 int  svipc_shm_write(long key, char *id, slot_array *a, int publish);
-int  svipc_shm_read(long key, char *id, slot_array *a, int subscribe);
+int  svipc_shm_read(long key, char *id, slot_array *a, float subscribe);
 int  svipc_shm_free(long key, char* id);
 int  svipc_shm_cleanup(long key);
